requirements.txt
python-status>=1.0
requests>=2.18
Flask>=0.12
pygame>=1.9.4
pygame-menu>=1.96.1
pyautogui>=0.9.42
jsonpickle>=1.0
cryptography>=2.0
